+++ src/indexing.c
@@ -16,7 +16,7 @@ void create_product_index(void)
 
     while (fread(&buffer, sizeof(ProductEntry), 1, products_file))
     {
-        long long address = ftell(products_file) - sizeof(ProductEntry);
+        long address = ftell(products_file) - sizeof(ProductEntry);
 
         if (record_count % MAX_INDEX_SPAN == 0)
         {
